app_server/requirements.txt
fastapi
uvicorn[standard]
aerospike
jsonpickle
python-snappy
kafka-python